+++ app/src/main/java/com/example/absolute_cinema/domain/model/MoviePoster.kt
@@ -9,5 +9,5 @@ data class MoviePoster(
     val filmCritics: Double?,
     val imdb: Double?,
     val kp: Double?,
-    val russianFilmCritics: Int?
+    val russianFilmCritics: Double?
 )
